pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "alignbooks-sdk"
version = "0.1.0"
description = "Unofficial Python SDK for AlignBooks Accounting/ERP API"
readme = "README.md"
authors = [
    {name = "OpenClaw GenAuto", email = "genautollp@gmail.com"},
]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.8"
dependencies = [
    "requests>=2.25.0",
    "pycryptodome>=3.10.0",
]

[project.urls]
Homepage = "https://github.com/genautollp/alignbooks-sdk"
Repository = "https://github.com/genautollp/alignbooks-sdk"
